requirements.txt
psycopg2-binary
python-dotenv
requests
camoufox
fastapi
uvicorn[standard]
openai
python-multipart
